versions the API and places it behind a URL_PREFIX
Running the "whatson" service behind the API Gateway means we have to
use path-based routing.
eg.
    All requests to /whats-on --> <whats-on-lambda>

That means that when the requests arrive at the Lambda function they
will contain the "/whats-on" string at the beginning of the request
path.

This PR provides a mechanism to handle the path prefix in a single
location.

I've also moved some files around a bit to support a versioned API,
which is a pattern we use for all services behind the API gateway.