remove previously extracted variable
Decided that pulling it out was unneccesary. If, in the future, it's
being used somewhere else, it should be pulled out. But for now, I think
that with the reliability of the input, the readability benefits of the
f-string, and the limited scope make it ok to have this "magic string"
in this function.